+++ spec/presentation/slide/timing_spec.rb
@@ -15,7 +15,13 @@ describe 'My behaviour' do
       it 'condition_no_event' do
         pptx = OoxmlParser::PptxParser.parse_pptx('spec/presentation/slide/timing/time_node/condition/condition_no_event.pptx')
         expect(pptx.slides[1].timing.time_node_list.first.common_time_node.children.first
-                   .common_time_node.children.first.common_time_node.children.first.common_time_node.start_conditions.first.event).to be_nil
+                   .common_time_node.children.first.common_time_node.children.first.common_time_node.start_conditions.list.first.event).to be_nil
+      end
+
+      it 'condition_list_check' do
+        pptx = OoxmlParser::PptxParser.parse_pptx('spec/presentation/slide/timing/time_node/condition/condition_list_check.pptx')
+        expect(pptx.slides.first.timing.time_node_list.first.common_time_node.children.first
+                   .common_time_node.children.first.common_time_node.start_conditions.list.first.delay).to eq(:indefinite)
       end
     end
   end